+++ actix-server/src/worker.rs
@@ -17,11 +17,11 @@ use actix_rt::{
     Arbiter, ArbiterHandle, System,
 };
 use futures_core::{future::LocalBoxFuture, ready};
-use log::{error, info, trace};
 use tokio::sync::{
     mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender},
     oneshot,
 };
+use tracing::{error, info, trace};
 
 use crate::{
     service::{BoxedServerService, InternalServiceFactory},
@@ -383,7 +383,7 @@ impl ServerWorker {
                             worker_stopped_rx.await.unwrap();
                         };
 
-                        #[cfg(all(target_os = "linux", feature = "io-uring"))]
+                        #[cfg(all(target_os = "linux", feature = "experimental-io-uring"))]
                         {
                             // TODO: pass max blocking thread config when tokio-uring enable configuration
                             // on building runtime.
@@ -391,7 +391,10 @@ impl ServerWorker {
                             tokio_uring::start(worker_fut);
                         }
 
-                        #[cfg(not(all(target_os = "linux", feature = "io-uring")))]
+                        #[cfg(not(all(
+                            target_os = "linux",
+                            feature = "experimental-io-uring"
+                        )))]
                         {
                             let rt = tokio::runtime::Builder::new_current_thread()
                                 .enable_all()
@@ -407,7 +410,7 @@ impl ServerWorker {
 
             // with actix system
             (Some(_sys), _) => {
-                #[cfg(all(target_os = "linux", feature = "io-uring"))]
+                #[cfg(all(target_os = "linux", feature = "experimental-io-uring"))]
                 let arbiter = {
                     // TODO: pass max blocking thread config when tokio-uring enable configuration
                     // on building runtime.
@@ -415,7 +418,7 @@ impl ServerWorker {
                     Arbiter::new()
                 };
 
-                #[cfg(not(all(target_os = "linux", feature = "io-uring")))]
+                #[cfg(not(all(target_os = "linux", feature = "experimental-io-uring")))]
                 let arbiter = {
                     Arbiter::with_tokio_rt(move || {
                         tokio::runtime::Builder::new_current_thread()